Adds an iframe and sets its headers to avoid cross-origin isolation.
Interim solution. some cleanup required.
