+++ package.json
@@ -29,6 +29,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-i18next": "^11.18.6",
+    "react-iframe": "^1.8.5",
     "react-modal": "^3.15.1",
     "react-router-dom": "^6.4.1",
     "react-scripts": "5.0.1",
