+++ public/signin.html
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html>
+<head>
+  <meta charset="utf-8" />
+  <title>SIWE Quickstart</title>
+</head>
+
+<body>
+  <div><button id='connectWalletBtn'>Connect wallet</button></div>
+  <div><button id='siweBtn'>Sign-in with Ethereum</button></div>
+</body>
+</html>
