+++ src/App.tsx
@@ -5,6 +5,7 @@ import {
   SigninPage,
   EntropyInputPage,
   DoubleSignPage,
+  DoubleSignIframePage,
   LobbyFullPage,
   LobbyPage,
   ContributingPage,
@@ -62,6 +63,14 @@ function App() {
               element={
                 <RequireAuth>
                   <DoubleSignPage />
+                </RequireAuth>
+              }
+            />
+            <Route
+              path={ROUTES.DOUBLE_SIGN_IFRAME}
+              element={
+                <RequireAuth>
+                  <DoubleSignIframePage />
                 </RequireAuth>
               }
             />
