+++ src/pages/index.ts
@@ -8,6 +8,7 @@ import ContributingPage from './contributing'
 import CompletePage from './complete'
 import RecordPage from './record'
 import DoubleSignPage from './doubleSign'
+import DoubleSignIframePage from './doubleSignIframe'
 import SigninRedirectPage from './signinRedirect'
 import FaqPage from './faq'
 
@@ -21,6 +22,7 @@ export {
   ContributingPage,
   CompletePage,
   DoubleSignPage,
+  DoubleSignIframePage,
   RecordPage,
   SigninRedirectPage,
   FaqPage
